requirements.txt
Babel==2.4.0
bcrypt==3.1.3
blinker==1.4
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.2
Flask-User==0.6.11
Flask-WTF==0.14.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
paho-mqtt==1.3.0
passlib==1.7.1
pycparser==2.17
pycryptodome==3.4.6
pytz==2017.2
requests==2.18.1
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.11
urllib3==1.21.1
Werkzeug==0.12.2
WTForms==2.1
